# Improve Matrix RREF Implementation