## What's new ?

For decoding coded pieces, coding coefficient matrix is converted into **R**educed **R**ow **E**chelon **F**orm, everytime new piece arrives --- I attempt to improve `rref` implementation by making mutation in-place ( reduction in memory allocation ), resulting into **>=2x** speed up.

![out](https://user-images.githubusercontent.com/45074836/128623992-170bb67c-847a-43f5-b763-92cd616385a3.png)

I do benchmark comparison of old & new implementations 

```bash
$ benchcmp new.txt old.txt 

benchmark                           old ns/op      new ns/op      delta
BenchmarkMatrixRref/2x2-8           264            270            +2.35%
BenchmarkMatrixRref/4x4-8           568            869            +53.02%
BenchmarkMatrixRref/8x8-8           2269           3909           +72.28%
BenchmarkMatrixRref/16x16-8         13208          22458          +70.03%
BenchmarkMatrixRref/32x32-8         92794          129114         +39.14%
BenchmarkMatrixRref/64x64-8         681918         848751         +24.47%
BenchmarkMatrixRref/128x128-8       5568454        6064303        +8.90%
BenchmarkMatrixRref/256x256-8       43729678       44994637       +2.89%
BenchmarkMatrixRref/512x512-8       333661594      344285636      +3.18%
BenchmarkMatrixRref/1024x1024-8     2654513014     2825216190     +6.43%

benchmark                           old MB/s     new MB/s     speedup
BenchmarkMatrixRref/2x2-8           30.33        14.81        0.49x
BenchmarkMatrixRref/4x4-8           56.33        18.41        0.33x
BenchmarkMatrixRref/8x8-8           56.41        16.37        0.29x
BenchmarkMatrixRref/16x16-8         38.76        11.40        0.29x
BenchmarkMatrixRref/32x32-8         22.07        7.93         0.36x
BenchmarkMatrixRref/64x64-8         12.01        4.83         0.40x
BenchmarkMatrixRref/128x128-8       5.88         2.70         0.46x
BenchmarkMatrixRref/256x256-8       3.00         1.46         0.49x
BenchmarkMatrixRref/512x512-8       1.57         0.76         0.48x
BenchmarkMatrixRref/1024x1024-8     0.79         0.37         0.47x

benchmark                           old allocs     new allocs     delta
BenchmarkMatrixRref/2x2-8           6              8              +33.33%
BenchmarkMatrixRref/4x4-8           10             32             +220.00%
BenchmarkMatrixRref/8x8-8           18             128            +611.11%
BenchmarkMatrixRref/16x16-8         34             510            +1400.00%
BenchmarkMatrixRref/32x32-8         66             2040           +2990.91%
BenchmarkMatrixRref/64x64-8         130            8160           +6176.92%
BenchmarkMatrixRref/128x128-8       258            32639          +12550.78%
BenchmarkMatrixRref/256x256-8       514            130567         +25302.14%
BenchmarkMatrixRref/512x512-8       1026           522223         +50798.93%
BenchmarkMatrixRref/1024x1024-8     2050           2089044        +101804.59%

benchmark                           old bytes     new bytes      delta
BenchmarkMatrixRref/2x2-8           104           63             -39.42%
BenchmarkMatrixRref/4x4-8           224           223            -0.45%
BenchmarkMatrixRref/8x8-8           512           1211           +136.52%
BenchmarkMatrixRref/16x16-8         1280          8543           +567.42%
BenchmarkMatrixRref/32x32-8         3584          66043          +1742.72%
BenchmarkMatrixRref/64x64-8         11264         523754         +4549.80%
BenchmarkMatrixRref/128x128-8       38912         4180774        +10644.18%
BenchmarkMatrixRref/256x256-8       143360        33430985       +23219.60%
BenchmarkMatrixRref/512x512-8       548864        267387738      +48616.57%
BenchmarkMatrixRref/1024x1024-8     2146304       2139164576     +99567.36%
```